+++ CHANGELOG.md
@@ -22,6 +22,7 @@
   * Osm::FlexiRecord
     * Add add\_field method to add a field to the record in OSM
     * Add update\_field method to rename a field in OSM
+    * Add delete\_field method to delete a field from OSM
 
 ## Version 0.1.11